+++ src/tung_websocket/notifier.rs
@@ -99,9 +99,15 @@ impl Notifier
 
 						Ok(()) =>
 						{
-							// note we can only get here if the queue isn't empty, so unwrap
+							// note we can only get here if the queue isn't empty, shouldn't happen
 							//
-							if let Err(_e) = pharos.as_mut().start_send( self.events.pop_front().expect( "pop queued event." ) )
+							let event = match self.events.pop_front()
+							{
+								Some(e) => e,
+								None => return Poll::Ready(Err(())),
+							};
+
+							if let Err(_e) = pharos.as_mut().start_send( event )
 							{
 								self.state = State::Closed;
 
@@ -212,7 +218,7 @@ mod tests
 			assert_eq!( State::Ready, not.state );
 
 
-		not.queue( WsEvent::Ping( vec![ 1, 2, 3] ) );
+		not.queue( WsEvent::Ping( vec![ 1, 2, 3].into() ) );
 
 			assert_eq!( State::Pending, not.state );
 
@@ -247,8 +253,8 @@ mod tests
 
 		// Queue 2 so the channel gives back pressure.
 		//
-		not.queue( WsEvent::Ping( vec![ 1, 2, 3] ) );
-		not.queue( WsEvent::Ping( vec![ 1, 2, 3] ) );
+		not.queue( WsEvent::Ping( vec![ 1, 2, 3].into() ) );
+		not.queue( WsEvent::Ping( vec![ 1, 2, 3].into() ) );
 
 			assert_eq!( State::Pending, not.state        );
 			assert_eq!(              2, not.events.len() );
@@ -300,7 +306,7 @@ mod tests
 			assert_eq!( 0, not.events.len() );
 
 
-		not.queue( WsEvent::Ping( vec![ 1, 2, 3] ) );
+		not.queue( WsEvent::Ping( vec![ 1, 2, 3].into() ) );
 
 			assert_eq!( 1, not.events.len() );