+++ tests/protocol_error.rs
@@ -3,7 +3,7 @@
 use
 {
 	ws_stream_tungstenite :: { *                                                    } ,
-	futures               :: { StreamExt, SinkExt, future::join                     } ,
+	futures               :: { StreamExt, future::join                              } ,
 	asynchronous_codec    :: { LinesCodec, Framed                                   } ,
 	tokio                 :: { net::{ TcpListener }                                 } ,
 	async_tungstenite     :: { accept_async, tokio::{ connect_async, TokioAdapter } } ,
@@ -48,7 +48,7 @@ async fn protocol_error()
 		let url             = Url::parse( "ws://127.0.0.1:3016" ).unwrap();
 		let (mut socket, _) = connect_async( url ).await.expect( "ws handshake" );
 
-		socket.send( tungstenite::Message::Ping( vec![1;126] ) ).await.expect( "send ping" );
+		socket.send( tungstenite::Message::Ping( vec![1;126].into() ) ).await.expect( "send ping" );
 
 		socket.close( None ).await.expect( "close client end" );